Coding Days: level type changed to int
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2761 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/finance/application/BlackScholesHullWhiteSolver.cpp
@@ -80,7 +80,7 @@ BlackScholesHullWhiteSolver::~BlackScholesHullWhiteSolver()
 	}
 }
 
-void BlackScholesHullWhiteSolver::constructGrid(BoundingBox& BoundingBox, size_t level)
+void BlackScholesHullWhiteSolver::constructGrid(BoundingBox& BoundingBox, int level)
 {
 	this->dim = BoundingBox.getDimensions();
 	this->levels = level;
